requirements/dev.txt
sacred>=0.8.4
pymongo>=4.6.1
pytorch-lightning>=2.1.3
hydra-core>=1.3.2
tqdm>=4.66.1
rich>=13.7.0
# Testing and linting
pytest>=8.0.0
pytest-benchmark>=4.0.0
pytest-ruff>=0.3.0
pre-commit>=3.5.0
black>=24.0.0
ruff>=0.3.0
pylint>=3.0.3
